Move implementations to cpp
